+++ src/chains/definitions/optimismGoerli.ts
@@ -36,6 +36,11 @@ export const optimismGoerli = /*#__PURE__*/ defineChain({
         address: '0x5b47E1A08Ea6d985D6649300584e6722Ec4B1383',
       },
     },
+    l1StandardBridge: {
+      [sourceId]: {
+        address: '0x636Af16bf2f682dD3109e60102b8E1A089FedAa8',
+      },
+    },
   },
   testnet: true,
   sourceId,