+++ src/chains/definitions/xdcTestnet.ts
@@ -17,4 +17,10 @@ export const xdcTestnet = /*#__PURE__*/ defineChain({
       url: 'https://apothem.blocksscan.io',
     },
   },
+  contracts: {
+    multicall3: {
+      address: '0xca11bde05977b3631167028862be2a173976ca11',
+      blockCreated: 59765389,
+    },
+  },
 })