+++ src/chains/definitions/zoraSepolia.ts
@@ -42,6 +42,11 @@ export const zoraSepolia = /*#__PURE__*/ defineChain({
         address: '0xeffE2C6cA9Ab797D418f0D91eA60807713f3536f',
       },
     },
+    l1StandardBridge: {
+      [sourceId]: {
+        address: '0x5376f1D543dcbB5BD416c56C189e4cB7399fCcCB',
+      },
+    },
   },
   sourceId,
   testnet: true,