+++ src/chains/zksync/formatters.test-d.ts
@@ -149,7 +149,7 @@ describe('smoke', () => {
     })
 
     expectTypeOf(transaction.type).toEqualTypeOf<
-      'legacy' | 'eip2930' | 'eip1559' | 'eip712' | 'priority'
+      'legacy' | 'eip2930' | 'eip1559' | 'eip4844' | 'eip712' | 'priority'
     >()
   })