+++ src/errors/address.test.ts
@@ -10,6 +10,9 @@ test('InvalidAddressError', () => {
   ).toMatchInlineSnapshot(`
     [InvalidAddressError: Address "0xa5cc3c03994db5b0d9a5eEdD10Cabab0813678az" is invalid.
 
+    - Address must be a hex value of 20 bytes (40 hex characters).
+    - Address must match its checksum counterpart.
+
     Version: viem@1.0.2]
   `)
 })