+++ src/index.ts
@@ -852,6 +852,12 @@ export {
   UrlRequiredError,
   type UrlRequiredErrorType,
 } from './errors/transport.js'
+export {
+  AccountStateConflictError,
+  type AccountStateConflictErrorType,
+  StateAssignmentConflictError,
+  type StateAssignmentConflictErrorType,
+} from './errors/stateOverride.js'
 export type {
   AbiItem,
   ExtractAbiFunctionForArgs,
@@ -872,6 +878,7 @@ export type {
   TransactionBase,
   TransactionEIP1559,
   TransactionEIP2930,
+  TransactionEIP4844,
   TransactionLegacy,
   TransactionReceipt,
   TransactionRequest,
@@ -942,6 +949,7 @@ export type {
   FeeHistory,
   FeeValues,
   FeeValuesEIP1559,
+  FeeValuesEIP4844,
   FeeValuesLegacy,
   FeeValuesType,
 } from './types/fee.js'
@@ -971,8 +979,15 @@ export type {
   RpcUncle,
   Status,
   RpcProof,
+  RpcAccountStateOverride,
+  RpcStateOverride,
+  RpcStateMapping,
 } from './types/rpc.js'
 export type { Withdrawal } from './types/withdrawal.js'
+export type {
+  StateMapping,
+  StateOverride,
+} from './types/stateOverride.js'
 export { labelhash, type LabelhashErrorType } from './utils/ens/labelhash.js'
 export { namehash, type NamehashErrorType } from './utils/ens/namehash.js'
 export {