+++ src/utils/formatters/transaction.ts
@@ -37,6 +37,7 @@ export const transactionType = {
   '0x0': 'legacy',
   '0x1': 'eip2930',
   '0x2': 'eip1559',
+  '0x3': 'eip4844',
 } as const
 
 export type FormatTransactionErrorType = ErrorType
@@ -51,6 +52,9 @@ export function formatTransaction(transaction: Partial<RpcTransaction>) {
     chainId: transaction.chainId ? hexToNumber(transaction.chainId) : undefined,
     gas: transaction.gas ? BigInt(transaction.gas) : undefined,
     gasPrice: transaction.gasPrice ? BigInt(transaction.gasPrice) : undefined,
+    maxFeePerBlobGas: transaction.maxFeePerBlobGas
+      ? BigInt(transaction.maxFeePerBlobGas)
+      : undefined,
     maxFeePerGas: transaction.maxFeePerGas
       ? BigInt(transaction.maxFeePerGas)
       : undefined,
@@ -86,13 +90,18 @@ export function formatTransaction(transaction: Partial<RpcTransaction>) {
 
   if (transaction_.type === 'legacy') {
     delete transaction_.accessList
+    delete transaction_.maxFeePerBlobGas
     delete transaction_.maxFeePerGas
     delete transaction_.maxPriorityFeePerGas
     delete transaction_.yParity
   }
   if (transaction_.type === 'eip2930') {
+    delete transaction_.maxFeePerBlobGas
     delete transaction_.maxFeePerGas
     delete transaction_.maxPriorityFeePerGas
+  }
+  if (transaction_.type === 'eip1559') {
+    delete transaction_.maxFeePerBlobGas
   }
   return transaction_
 }