+++ src/utils/formatters/transactionRequest.ts
@@ -20,6 +20,7 @@ export const rpcTransactionType = {
   legacy: '0x0',
   eip2930: '0x1',
   eip1559: '0x2',
+  eip4844: '0x3',
 } as const
 
 export type FormatTransactionRequestErrorType = ErrorType